requirements.txt
autograd==1.3
Cython==0.29.28
h5py==3.6.0
matplotlib==3.4.3
networkx==2.5
nose==1.3.7
numba==0.54.1
pandas==1.3.4
pybasicbayes==0.2.2
pylds==0.0.5
PyYAML==6.0
ruamel.base==1.0.0
ruamel.yaml==0.17.21
scikit_image==0.19.2
scikit_learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
torchmetrics==0.7.2
tqdm==4.62.3
umap==0.1.1
umap_learn==0.5.2
opencv-python==4.6.0.66
einops
opt_einsum
plumbum
